+++ src/lightning/pytorch/core/module.py
@@ -619,7 +619,8 @@ class LightningModule(
     ) -> Union[Tensor, Dict, List, Tuple]:
         r"""Gather tensors or collections of tensors from multiple processes.
 
-        This method needs to be called on all processes. Failing to do so will cause your program to stall forever.
+        This method needs to be called on all processes and the tensors need to have the same shape across all
+        processes, otherwise your program will stall forever.
 
         Args:
             data: int, float, tensor of shape (batch, ...), or a (possibly nested) collection thereof.