+++ tests/tests_fabric/test_fabric.py
@@ -637,6 +637,11 @@ def test_no_backward_sync():
     with fabric.no_backward_sync(model):
         pass
     fabric._strategy._backward_sync_control.no_backward_sync.assert_not_called()
+    # same for XLA
+    fabric._strategy = Mock(spec=XLAStrategy, _backward_sync_control=MagicMock())
+    with fabric.no_backward_sync(model):
+        pass
+    fabric._strategy._backward_sync_control.no_backward_sync.assert_not_called()
 
     # pretend that the strategy supports skipping backward sync
     fabric._strategy = Mock(_backward_sync_control=MagicMock())
@@ -644,7 +649,7 @@ def test_no_backward_sync():
     with fabric.no_backward_sync(model, enabled=False):
         pass
     fabric._strategy._backward_sync_control.no_backward_sync.assert_not_called()
-    # when enabld, the wrapped module gets passed down
+    # when enabled, the wrapped module gets passed down
     with fabric.no_backward_sync(model):
         pass
     fabric._strategy._backward_sync_control.no_backward_sync.assert_called_once_with(model._forward_module)
@@ -739,7 +744,7 @@ def test_init_module_context(monkeypatch):
     fabric._strategy = strategy
     with fabric.init_module():
         pass
-    strategy.module_init_context.assert_called_once()
+    strategy.module_init_context.assert_called_once_with(empty_init=None)
     strategy.module_init_context.reset_mock()
 
     # Pretend we are using PyTorch < 2.0
@@ -815,6 +820,29 @@ def test_call():
     with pytest.warns(UserWarning, match="Skipping the callback `Mock.not_a_method`"):
         fabric.call("not_a_method")
     assert not callback1.mock_calls
+
+
+def test_special_callbacks():
+    """Tests special callbacks that have hooks for internal Fabric events."""
+
+    class SpecialCallback:
+        def on_after_optimizer_step(self, strategy, optimizer):
+            pass
+
+        def on_after_setup(self, fabric, module):
+            pass
+
+    callback = Mock(wraps=SpecialCallback())
+    fabric = Fabric(accelerator="cpu", callbacks=[callback])
+
+    model = torch.nn.Linear(2, 2)
+    optimizer = torch.optim.SGD(model.parameters(), lr=0.1)
+    fabric_model, fabric_optimizer = fabric.setup(model, optimizer)
+    callback.on_after_setup.assert_called_once_with(fabric=fabric, module=fabric_model)
+
+    model(torch.randn(2, 2)).sum().backward()
+    fabric_optimizer.step()
+    callback.on_after_optimizer_step.assert_called_once_with(strategy=fabric._strategy, optimizer=optimizer)
 
 
 def test_loggers_input():
@@ -913,7 +941,47 @@ def test_save_wrapped_objects(setup, tmp_path):
     state = {"model": model, "optimizer": optimizer, "anything": anything}
     expected = {"model": unwrapped_model, "optimizer": unwrapped_optimizer, "anything": anything}
     fabric.save(tmp_path, state)
-    save_checkpoint_mock.assert_called_with(state=expected, path=tmp_path)
+    save_checkpoint_mock.assert_called_with(state=expected, path=tmp_path, filter=None)
+
+
+def test_save_filter(tmp_path):
+    fabric = Fabric(devices=1)
+    checkpoint_io_mock = Mock()
+    fabric.strategy.checkpoint_io = checkpoint_io_mock
+
+    model = BoringModel()
+    optimizer = torch.optim.Adam(model.parameters())
+
+    anything = {"cocofruit": 1}
+    state = {"model": model, "optimizer": optimizer, "anything": anything, "foo": 1}
+    save_path = tmp_path / "foo.pth"
+
+    # filter all dicts
+    filter = {k: lambda k, v: False for k in state}
+    fabric.save(save_path, state, filter=filter)
+    checkpoint_io_mock.save_checkpoint.assert_called_with(checkpoint={"foo": 1}, path=save_path, storage_options=None)
+
+    # bad filters
+    with pytest.raises(TypeError, match="should be a dict"):
+        fabric.save(save_path, state, filter="foo")
+    with pytest.raises(TypeError, match="callable, given 'foo"):
+        fabric.save(save_path, state, filter={"model": "foo"})
+    with pytest.raises(ValueError, match="keys {'asd'} are not present in the state keys"):
+        fabric.save(save_path, state, filter={"asd": lambda k, v: True})
+
+    # subset
+    checkpoint_io_mock.reset_mock()
+    filter = {
+        "model": lambda k, v: "weight" in k,
+        "anything": lambda k, v: isinstance(v, int),
+        "optimizer": lambda k, v: "param_groups" in k,
+    }
+    fabric.save(save_path, state, filter=filter)
+    checkpoint_io_mock.save_checkpoint.assert_called_with(
+        checkpoint={"model": {"layer.weight": ANY}, "optimizer": {"param_groups": ANY}, "anything": anything, "foo": 1},
+        path=save_path,
+        storage_options=None,
+    )
 
 
 @pytest.mark.parametrize("setup", [True, False])
@@ -923,7 +991,7 @@ def test_load_wrapped_objects(setup, tmp_path):
 
     expected_remainder = {"extra": "data"}
 
-    def mocked_load_checkpoint(path, state):
+    def mocked_load_checkpoint(path, state, strict):
         assert not isinstance(state["model"], _FabricModule)
         assert not isinstance(state["optimizer"], _FabricOptimizer)
         state.update({"int": 5, "dict": {"x": 1}})
@@ -1008,6 +1076,26 @@ def test_all_reduce():
     # dict
     fabric.all_reduce({"a": torch.tensor(4), "b": [torch.tensor(5)], "c": "string"})
     fabric._strategy.all_reduce.assert_has_calls([call(torch.tensor(4), **defaults), call(torch.tensor(5), **defaults)])
+
+
+def test_rank_zero_first():
+    """Test that rank 0 completes first before all other processes can execute under `.rank_zero_first()`."""
+
+    def record_calls_for_rank(rank):
+        call_order = []
+
+        fabric = Fabric()
+        fabric._strategy = Mock(global_rank=rank)
+        fabric.barrier = Mock(side_effect=lambda *_: call_order.append("barrier"))
+        target = Mock(run=Mock(side_effect=lambda *_: call_order.append("run")))
+
+        with fabric.rank_zero_first():
+            target.run()
+
+        return call_order
+
+    assert record_calls_for_rank(0) == ["run", "barrier", "barrier"]
+    assert record_calls_for_rank(1) == ["barrier", "run", "barrier"]
 
 
 @pytest.mark.parametrize(("clip_val", "max_norm"), [(1e-3, None), (None, 1)])